% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberOfRealRoots.R
\name{numberOfRealRoots}
\alias{numberOfRealRoots}
\title{Number of real roots}
\usage{
numberOfRealRoots(qspray)
}
\arguments{
\item{qspray}{a univariate \code{qspray} polynomial}
}
\value{
An integer, the number of real roots of the polynomial.
}
\description{
Number of distinct real roots of a univariate polynomial.
}
\note{
The roots are not counted with their multiplicity.
}
\examples{
library(resultant)
x <- qlone(1)
P <- 2*x^4 + x^3 - 3*x^2 - x + 1
numberOfRealRoots(P)
}
