% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultant.R
\name{resultant}
\alias{resultant}
\title{Resultant of two polynomials}
\usage{
resultant(qspray1, qspray2)
}
\arguments{
\item{qspray1, qspray2}{two \code{qspray} polynomials with at more two
variables}
}
\value{
For univariate polynomials, this returns a fraction given as a
  string. For bivariate polynomials, this returns a univariate
  \code{qspray} polynomial.
}
\description{
Resultant of two polynomials with rational coefficients.
}
\examples{
library(resultant)
library(qspray)
x <- qlone(1)
y <- qlone(2)
f <- x^4 - x^3 + x^2 - 2*x*y^2 + y^4
g <- x - 2*y^2
resultant(f, g)
}
