% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subresultants.R
\name{subresultants}
\alias{subresultants}
\title{Subresultants of two polynomials}
\usage{
subresultants(qspray1, qspray2, var = 1)
}
\arguments{
\item{qspray1, qspray2}{two \code{qspray} polynomials with at more nine
variables}

\item{var}{integer indicating with respect to which variable the resultant
is desired (e.g. \code{1} for \code{x} and \code{2} for \code{y})}
}
\value{
For univariate polynomials, this returns a vector of fractions
  given as strings.
  For bivariate polynomials, this returns a list of univariate
  \code{qspray} polynomials. And so on.
}
\description{
Subresultants of two polynomials with rational coefficients.
}
\examples{
library(resultant)
library(qspray)
x <- qlone(1)
y <- qlone(2)
p <- x^2 * y * (y^2 - 5*x + 6)
q <- x^2 * y * (3*y + 2)
SRx <- subresultants(p, q, var = 1) # should be 0, 0, non-zero, ...
# lapply(SRx, function(s) prettyQspray(s, "y"))
SRy <- subresultants(p, q, var = 2) # should be 0, non-zero, ...
# lapply(SRy, function(s) prettyQspray(s, "x"))
}
